continue making change to account for new ux
